clean code: Move Netex index interfaces/api classes into the index.api package. This prevent cyclic dependencies and allow for a cleaner dependency graph.
