+++ src/main/java/org/opentripplanner/netex/index/api/ReadOnlyHierarchicalVersionMapById.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.index.hierarchy;
+package org.opentripplanner.netex.index.api;
 
 import java.util.Collection;
