+++ src/main/java/org/opentripplanner/netex/mapping/FlexStopLocationMapper.java
@@ -1,7 +1,6 @@
-package org.opentripplanner.netex.loader.mapping;
+package org.opentripplanner.netex.mapping;
 
 import org.opentripplanner.model.FlexStopLocation;
-import org.opentripplanner.netex.mapping.FeedScopedIdFactory;
 import org.rutebanken.netex.model.FlexibleArea;
 import org.rutebanken.netex.model.FlexibleStopPlace;
 import org.slf4j.Logger;
